Added simple plane, player sphere, simple player controller script, and camera controller script
Initial project set up. Camera controller script contains commented out code. Might come back to it later to add a way for the camera to follow the ball on all axises while still allowing the sphere to roll. Otherwise, will either keep it this way with following the player's transform at an offest, or will simply parent the camera to the player and add constraints to lock the rotation in the inspector. Will come back to this script.